user thongbao

+++ database/migrations/2020_06_27_155038_create_chi_tiet_hoa_dons_table.php
@@ -19,7 +19,7 @@ class CreateChiTietHoaDonsTable extends Migration
             $table->Integer('can_ho_id');
             $table->Integer('dich_vu_id');
             $table->Float('so_luong');
-            $table->double('thanh_tien');
+            $table->Float('thanh_tien');
             $table->softDeletes();
             $table->timestamps();    
         });
