+++ database/migrations/2020_07_24_044545_create_dich_vus_table.php
@@ -17,7 +17,7 @@ class CreateDichVusTable extends Migration
             $table->increments('id');
             $table->string('ten_dich_vu');
             $table->string('don_vi');
-            $table->double('phi_dv');
+            $table->Float('phi_dv');
             $table->softDeletes();
             $table->timestamps();
         });
